+++ app/src/main/java/com/accenture/rishikeshpoorun/moFaim/DataLayer/Entities/User.java
@@ -10,9 +10,9 @@ import android.arch.persistence.room.PrimaryKey;
 public class User {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "USER_ID")
-    private Integer userId;
+    private Long userId;
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -36,7 +36,7 @@ public class User {
     }
 
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
@@ -77,4 +77,6 @@ public class User {
                 ", password='" + password + '\'' +
                 '}';
     }
+
+
 }
